Rewrite to more modular CLI
