+++ Cargo.toml
@@ -48,6 +48,8 @@ rand = "0.8.5"
 typenum = "1.17.0"
 log = { version = "0.4.22" }
 clap = { version = "4.5.18", features = ["derive"] }
+clap-verbosity-flag = "2.2.1"
+clap_mangen = "0.2.23"
 serde = { version = "1.0.210", features = ["derive"] }
 arbitrary = { version = "1.3.2", features = ["derive"] }
 anyhow = { version = "1.0.89", features = ["backtrace", "std"] }
