+++ rosenpass/src/lib.rs
@@ -1,8 +1,11 @@
 #[cfg(feature = "experiment_api")]
 pub mod api;
 pub mod app_server;
+pub mod broker;
 pub mod cli;
+pub mod cmd;
 pub mod config;
+pub mod event_loop;
 pub mod hash_domains;
 pub mod msgs;
 pub mod protocol;
