+++ rosenpass/src/main.rs
@@ -1,13 +1,9 @@
 use clap::Parser;
-use log::error;
-use rosenpass::cli::CliArgs;
-use std::process::exit;
+use rosenpass::broker;
+use rosenpass::cli::{Cli, Commands};
+use rosenpass::cmd::Command;
 
-/// Catches errors, prints them through the logger, then exits
-pub fn main() {
-    // parse CLI arguments
-    let args = CliArgs::parse();
-
+pub fn main() -> anyhow::Result<()> {
     {
         use rosenpass_secret_memory as SM;
         #[cfg(feature = "experiment_memfd_secret")]
@@ -16,30 +12,21 @@ pub fn main() {
         SM::secret_policy_use_only_malloc_secrets();
     }
 
-    // init logging
-    {
-        let mut log_builder = env_logger::Builder::from_default_env(); // sets log level filter from environment (or defaults)
-        if let Some(level) = args.get_log_level() {
-            log::debug!("setting log level to {:?} (set via CLI parameter)", level);
-            log_builder.filter_level(level); // set log level filter from CLI args if available
-        }
-        log_builder.init();
+    let cli = Cli::parse();
 
-        // // check the effectiveness of the log level filter with the following lines:
-        // use log::{debug, error, info, trace, warn};
-        // trace!("trace dummy");
-        // debug!("debug dummy");
-        // info!("info dummy");
-        // warn!("warn dummy");
-        // error!("error dummy");
-    }
+    env_logger::Builder::new()
+        .filter_level(cli.verbose.log_level_filter())
+        .init();
 
-    let broker_interface = args.get_broker_interface();
-    match args.run(broker_interface, None) {
-        Ok(_) => {}
-        Err(e) => {
-            error!("{e:?}");
-            exit(1);
-        }
+    let broker_interface = broker::get_broker_interface(&cli);
+
+    match cli.command {
+        Commands::ExchangeConfig(exchangeconfig) => exchangeconfig.run(broker_interface, None),
+        Commands::Exchange(exchange) => exchange.run(broker_interface, None),
+        Commands::GenConfig(genconfig) => genconfig.run(broker_interface, None),
+        Commands::GenKeys(genkeys) => genkeys.run(broker_interface, None),
+        Commands::Keygen(keygen) => keygen.run(broker_interface, None),
+        Commands::Validate(validate) => validate.run(broker_interface, None),
+        Commands::Man(man) => man.run(broker_interface, None),
     }
 }
